fixed tests for add payment form
